move pillow out into its own module
