+++ fluff/__init__.py
@@ -1,17 +1,13 @@
 from collections import defaultdict
 import functools
 import logging
-from couchdbkit import ResourceNotFound
 from couchdbkit.ext.django import schema
 import datetime
 import sqlalchemy
 from dimagi.utils.parsing import json_format_date
-from dimagi.utils.read_only import ReadOnlyObject
 from fluff import exceptions
 from fluff.exceptions import EmitterValidationError
 from fluff.signals import BACKEND_SQL, BACKEND_COUCH
-from pillowtop.checkpoints.manager import get_default_django_checkpoint_for_legacy_pillow_class
-from pillowtop.listener import PythonPillow, PYTHONPILLOW_CHUNK_SIZE
 from .signals import indicator_document_updated
 import fluff.util
 
@@ -641,6 +637,7 @@ class IndicatorDocument(schema.Document):
 
     @classmethod
     def pillow(cls):
+        from .pillow import FluffPillow
         wrapper = cls.wrapper or cls.document_class
         doc_type = cls.document_class._doc_type
         extra_args = dict(doc_type=doc_type)
@@ -687,116 +684,6 @@ class IndicatorDocument(schema.Document):
 
     class Meta:
         app_label = 'fluff'
-
-
-class FluffPillow(PythonPillow):
-    document_filter = None
-    wrapper = None
-    indicator_class = IndicatorDocument
-    domains = None
-    doc_type = None
-    save_direct_to_sql = False
-    delete_filtered = False  # delete docs not matching filter
-
-    # see explanation in IndicatorDocument for how this is used
-    deleted_types = ()
-
-    def __init__(self, chunk_size=None, checkpoint=None, change_feed=None):
-        # explicitly check against None since we want to pass chunk_size=0 through
-        chunk_size = chunk_size if chunk_size is not None else PYTHONPILLOW_CHUNK_SIZE
-        # fluff pillows should default to SQL checkpoints
-        checkpoint = checkpoint or get_default_django_checkpoint_for_legacy_pillow_class(self.__class__)
-        super(FluffPillow, self).__init__(
-            chunk_size=chunk_size,
-            checkpoint=checkpoint,
-            change_feed=change_feed,
-        )
-
-    @classmethod
-    def get_sql_engine(cls):
-        engine = getattr(cls, '_engine', None)
-        if not engine:
-            import sqlalchemy
-            from django.conf import settings
-            engine = sqlalchemy.create_engine(settings.SQL_REPORTING_DATABASE_URL)
-            cls._engine = engine
-        return engine
-
-    def python_filter(self, change):
-        doc = change.document
-        assert self.domains
-        assert None not in self.domains
-        assert self.doc_type is not None
-        assert self.doc_type not in self.deleted_types
-        if doc.get('domain') in self.domains:
-            return self._is_doc_type_match(doc.get('doc_type')) or self._is_doc_type_deleted_match(doc.get('doc_type'))
-
-    @classmethod
-    def _get_base_name(cls):
-        # used in the name/checkpoint ID
-        return 'fluff'
-
-    def _is_doc_type_match(self, type):
-        return type == self.doc_type
-
-    def _is_doc_type_deleted_match(self, type):
-        return type in self.deleted_types
-
-    def change_transform(self, doc_dict):
-        delete = False
-
-        doc = self.wrapper.wrap(doc_dict)
-        doc = ReadOnlyObject(doc)
-
-        if self.document_filter and not self.document_filter.filter(doc):
-            if self.delete_filtered:
-                delete = True
-            else:
-                return None
-
-        indicator = _get_indicator_doc_from_class_and_id(self.indicator_class, doc.get_id)
-        if not self._is_doc_type_deleted_match(doc.doc_type):
-            indicator.calculate(doc)
-        else:
-            indicator['id'] = doc.get_id
-            delete = True
-
-        return {
-            'doc_dict': doc_dict,
-            'indicators': indicator,
-            'delete': delete
-        }
-
-    def change_transport(self, data):
-        indicators = data['indicators']
-
-        diff = indicators.diff(None)  # pass in None for old_doc to force diff with ALL indicators
-        if self.save_direct_to_sql:
-            if not data['delete']:
-                indicators.save_to_sql(diff, self.get_sql_engine())
-            else:
-                indicators.delete_from_sql(self.get_sql_engine())
-        else:
-            if not data['delete']:
-                indicators.save()
-            else:
-                indicators.delete()
-
-        backend = BACKEND_SQL if self.save_direct_to_sql else BACKEND_COUCH
-        indicator_document_updated.send(
-            sender=self,
-            doc=data['doc_dict'],
-            diff=diff,
-            backend=backend
-        )
-
-
-def _get_indicator_doc_from_class_and_id(indicator_class, doc_id):
-    indicator_id = '%s-%s' % (indicator_class.__name__, doc_id)
-    try:
-        return indicator_class.get(indicator_id)
-    except ResourceNotFound:
-        return indicator_class(_id=indicator_id)
 
 
 try:
